+++ tests/js/user-tests.js
@@ -237,6 +237,23 @@ module.exports = {
     TestCase.assertEqual(await collection.count({hello: "pineapple"}), 0);
   },
 
+  async testPush() {
+    let app = new Realm.App(appConfig);
+    let credentials = Realm.Credentials.anonymous();
+    let user = await app.logIn(credentials);
+
+    let push = user.push('gcm');
+
+    await push.deregister(); // deregister never registered not an error
+    await push.register("hello");
+    await push.register("hello"); // double register not an error
+    await push.deregister();
+    await push.deregister(); // double deregister not an error
+
+    const err = await TestCase.assertThrowsAsync(async() => await user.push('nonesuch').register('hello'))
+    TestCase.assertEqual(err.message, "service not found: 'nonesuch'");
+  },
+
   async testAllWithAnonymous() {
     let app = new Realm.App(appConfig);
     await logOutExistingUsers(app);